Remove Sampler class
Rationale: Sampler has the same interface as RandomVector, to sample stuff (getRealization/getSample),
so MCMC inherits RandomVector.

This makes PosteriorRandomVector useless.

Also RWMH now sets the description of the prior when getSample is called (that was my original goal).
