+++ TODO
@@ -6,3 +6,4 @@ Remove deprecated ResourceMap::[SG]AsNumericalScalar
 Remove deprecated SpecFunc::*NumericalScalar* in 1.11
 Remove deprecated NumericalScalar[Collection] in 1.11
 Remove deprecated NumericalComplex[Collection] in 1.11
+Remove deprecated PosteriorRandomVector in 1.11
