+++ python/src/MCMC.i
@@ -8,3 +8,9 @@
 
 %include openturns/MCMC.hxx
 namespace OT { %extend MCMC { MCMC(const MCMC & other) { return new OT::MCMC(other); } } }
+
+%pythoncode %{
+def PosteriorRandomVector(mcmc):
+    openturns.common.Log.Warn('class PosteriorRandomVector is deprecated')
+    return mcmc
+%}
