--- python/src/Sampler_doc.i.in
@@ -1,14 +1,0 @@
-// Sampler docstrings are defined in SamplerImplementation_doc.i.in
-
-%feature("docstring") OT::Sampler
-OT_Sampler_doc
-%feature("docstring") OT::Sampler::getDimension
-OT_Sampler_getDimension_doc
-%feature("docstring") OT::Sampler::getRealization
-OT_Sampler_getRealization_doc
-%feature("docstring") OT::Sampler::getSample
-OT_Sampler_getSample_doc
-%feature("docstring") OT::Sampler::getVerbose
-OT_Sampler_getVerbose_doc
-%feature("docstring") OT::Sampler::setVerbose
-OT_Sampler_setVerbose_doc
