--- python/test/t_PosteriorRandomVector_std.expout
@@ -1,13 +1,0 @@
-randomVector= class=PosteriorRandomVector sampler=class=RandomWalkMetropolisHastings name=Unnamed derived from class=MCMC prior=class=ComposedDistribution name=ComposedDistribution dimension=2 copula=class=IndependentCopula name=IndependentCopula dimension=2 marginal[0]=class=Normal name=Normal dimension=1 mean=class=Point name=Unnamed dimension=1 values=[25] sigma=class=Point name=Unnamed dimension=1 values=[0.1] correlationMatrix=class=CorrelationMatrix dimension=1 implementation=class=MatrixImplementation name=Unnamed rows=1 columns=1 values=[1] marginal[1]=class=Dirac name=Dirac dimension=1 point=class=Point name=Unnamed dimension=1 values=[2] conditional=class=Normal name=Normal dimension=1 mean=class=Point name=Unnamed dimension=1 values=[0] sigma=class=Point name=Unnamed dimension=1 values=[1] correlationMatrix=class=CorrelationMatrix dimension=1 implementation=class=MatrixImplementation name=Unnamed rows=1 columns=1 values=[1] model=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x0,x1,y0,y1] evaluationImplementation=class=ParametricEvaluation function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x0,x1,y0,y1] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] parameters positions=[] parameters=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] input positions=[0,1] gradientImplementation=class=ParametricGradient evaluation=class=ParametricEvaluation function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x0,x1,y0,y1] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] parameters positions=[] parameters=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] input positions=[0,1] hessianImplementation=class=ParametricHessian evaluation=class=ParametricEvaluation function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x0,x1,y0,y1] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0,y1] formulas=[x0,x1] parameters positions=[] parameters=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] input positions=[0,1] parameters=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=10 dimension=0 data=[[],[],[],[],[],[],[],[],[],[]] observations=class=Sample name=Normal implementation=class=SampleImplementation name=Normal size=10 dimension=1 description=[X0] data=[[31.7298],[29.4806],[30.4741],[32.4466],[28.3823],[31.4201],[30.574],[32.7247],[31.9728],[31.9518]] burnIn=1000 thinning=100 proposal=[class=Uniform name=Uniform dimension=1 a=-2 b=2,class=Uniform name=Uniform dimension=1 a=-2 b=2] calibrationStrategy=[class=CalibrationStrategyImplementation name=Unnamed,class=CalibrationStrategyImplementation name=Unnamed]
-dimension= 2
-realization= [25.1464,2]
-sample= 0 : [ 25.154   2      ]
-1 : [ 25.225   2      ]
-2 : [ 25.3727  2      ]
-3 : [ 24.9898  2      ]
-4 : [ 25.154   2      ]
-5 : [ 25.2007  2      ]
-6 : [ 25.0898  2      ]
-7 : [ 25.0216  2      ]
-8 : [ 25.0886  2      ]
-9 : [ 25.0441  2      ]
